split quic layer into subpackages
this commit should not introduce any functional changes
